compiler/internal/vm/static_core/plugins/ets/playground/backend/pyproject.toml
# Copyright (c) 2024-2026 Huawei Device Co., Ltd.
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "arkts-playground"
dynamic = ["version"]
dependencies = [
    "fastapi[standard]",
    "PyYaml",
    "pydantic_settings",
    "aiofiles",
    "pydantic",
    "structlog>=25.5.0",
    "asgi-correlation-id>=4.3.4",
    "prometheus-client>=0.23.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
]
requires-python = ">=3.10"


[build-system]
requires = ["setuptools", "setuptools_scm", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools_scm]
write_to = "src/arkts_playground/_version.py"
fallback_version = "0.0.7"
version_scheme = "release-branch-semver"

[project.scripts]
arkts-playground = "arkts_playground.main:main"
